+++ CHANGELOG.md
@@ -42,6 +42,8 @@ A pure Rust EtherCAT MainDevice supporting std and no_std environments.
 - [#330](https://github.com/ethercrab-rs/ethercrab/pull/330) (@theol0403) Stop re-reading
   potentially broken SyncManager types over CoE. Now the EEPROM is the source of truth for this
   information.
+- [#354](https://github.com/ethercrab-rs/ethercrab/pull/354) (@prrn) Don't enable Sync Managers with
+  a length of zero.
 
 ## [0.6.0] - 2025-03-29